% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_s2_tessel.R
\name{create_s2_tessel}
\alias{create_s2_tessel}
\title{create_s2_tessel}
\usage{
create_s2_tessel(lat_dist)
}
\arguments{
\item{lat_dist}{`numeric [°]` Desired istance between tessel centroids (in degrees)}
}
\value{
The function creates the datasets needed to perform tessellation of
 S2 acquisitions, with a given latitudinal spacing. Results are saved in
 data/S2tessels.
}
\description{
Create datasets to be used for tessellation of S2 acquisitions
 over Africa ----
}
\details{
DETAILS
}
\examples{
\dontrun{
create_s2_tessel(3)
}
}
\author{
Lorenzo Busetto, phD (2019)
}
